+++ mimic-bukkit/src/main/kotlin/internal/Compat.kt
@@ -1,0 +1,19 @@
+package ru.endlesscode.mimic.internal
+
+private val notSupportedCalls = mutableSetOf<String>()
+
+internal inline fun <T> callCompat(
+    key: String,
+    block: () -> T,
+    compat: () -> T,
+): T {
+    if (key !in notSupportedCalls) {
+        try {
+            return block()
+        } catch (_: NoSuchMethodError) {
+            notSupportedCalls.add(key)
+        }
+    }
+
+    return compat()
+}